+++ CHANGELOG.md
@@ -23,8 +23,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Moved the changelog check to a GitHub status check @faboweb
 * Moved to SDK v0.21.1 @faboweb
 * Using vue-jest for transpiling in tests for better code coverage output @faboweb
+* Default to gaia-7001 @okwme
 * Changed `yarn testnet` to `yarn start` @okwme
-* Automatically update wallet when a relevant transaction is detected @mappum
+* Default to gaia-7001 @okwme
 
 ### Fixed